Removed Player from Grid, playing with animation speed to move character around a little faster
0.15 seems to be the sweet spot. However, by disabling processing while moving I can still press the keys just too fast so I miss a movement.
I could play with having an input queue, so every key press gets queued and handled in order no matter what. This would be interesting for holding the keys down as that has fairly rapid repetition.

I'm not convinced I need a "grid" to track where things are. I believe using raycast or 2D collisions will work for this game.